moved RegistryStore to registry_store.py
