+++ sdk/python/tests/foo_registry_store.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 from feast.protos.feast.core.Registry_pb2 import Registry as RegistryProto
-from feast.registry import RegistryStore
+from feast.registry_store import RegistryStore
 from feast.repo_config import RegistryConfig
